이력서 수정시 이미지 업로드 완료

+++ src/main/java/com/example/leapit/resume/ResumeService.java
@@ -1,6 +1,7 @@
 package com.example.leapit.resume;
 
 import com.example.leapit._core.error.ex.Exception400;
+import com.example.leapit._core.error.ex.Exception403;
 import com.example.leapit._core.error.ex.Exception404;
 import com.example.leapit._core.error.ex.ExceptionApi404;
 import com.example.leapit.application.Application;
@@ -86,15 +87,15 @@ public class ResumeService {
         return listDTOs;
     }
 
-    public ResumeResponse.DetailDTO detail(int resumeId) { // TODO : Integer sessionUserId 매개변수 추가
+    public ResumeResponse.DetailDTO detail(Integer resumeId, Integer sessionUserId) { // TODO : Integer sessionUserId 매개변수 추가
         // 1. 이력서 존재 확인
         Resume resume =  resumeRepository.findByIdJoinUser(resumeId);
         if (resume == null) throw new Exception404("이력서가 존재하지 않습니다.");
 
         // 2. 이력서 주인 (권한) 확인
-//        if(!(resume.getUser().getId().equals(sessionUserId)) ) {
-//            throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
-//        }
+        if(!(resume.getUser().getId().equals(sessionUserId)) ) {
+            throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
+        }
 
         // 3. 이력서 DTO 조립
         List<ResumeTechStack> techStacks  = resumeTechStackRepository.findAllByResumeId(resumeId);
@@ -111,7 +112,7 @@ public class ResumeService {
     }
 
     @Transactional
-    public void delete(int resumeId) { // TODO : Integer sessionUserId 매개변수 추가
+    public void delete(Integer resumeId, Integer sessionUserId) { // TODO : Integer sessionUserId 매개변수 추가
         // 1. 이력서 존재 확인
         Resume resume = resumeRepository.findByIdJoinUser(resumeId);
         if (resume == null) throw new Exception404("이력서가 존재하지 않습니다.");
@@ -124,16 +125,15 @@ public class ResumeService {
         }
 
         // 2. 이력서 주인 (권한) 확인
-//        if(!(resume.getUser().getId().equals(sessionUserId)) ) {
-//            throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
-//        }
+        if(!(resume.getUser().getId().equals(sessionUserId)) ) {
+            throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
+        }
 
         // 3. 이력서 삭제
         resumeRepository.deleteById(resumeId);
     }
 
     public ResumeResponse.SaveDTO getSaveForm(Integer sessionUserId) {
-        if (sessionUserId == null) throw new Exception404("회원정보가 존재하지 않습니다");
         User user = userRepository.findById(sessionUserId);
         List<PositionType> positionTypes = positionTypeRepository.findAll();
         List<TechStack> techStacks = techStackRepository.findAll();
@@ -142,8 +142,6 @@ public class ResumeService {
 
     @Transactional
     public void save(ResumeRequest.SaveDTO saveDTO, MultipartFile photoUrlFile, User sessionUser) {
-        if (sessionUser == null) throw new ExceptionApi404("회원정보가 존재하지 않습니다");
-
         // 이미지
         String uploadDir = System.getProperty("user.dir") + "/upload/";
         try {
@@ -167,34 +165,30 @@ public class ResumeService {
     }
 
     @Transactional
-    public void update(Integer resumeId, ResumeRequest.UpdateDTO reqDTO){
+    public void update(Integer resumeId, Integer sessionUserId, ResumeRequest.UpdateDTO reqDTO, MultipartFile photoUrlFile){
         // 1. 이력서 존재 확인
         Resume resumePS = resumeRepository.findById(resumeId);
         if (resumePS == null) throw new Exception404("이력서가 존재하지 않습니다.");
 
-        // 2. (선택) 권한 확인
-        // if (!(resumePS.getUser().getId().equals(sessionUserId))) {
-        //     throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
-        // }
+        // 2. 권한 확인
+         if (!(resumePS.getUser().getId().equals(sessionUserId))) {
+             throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
+         }
 
         // 이미지
-//        String uploadDir = System.getProperty("user.dir") + "/upload/";
-//
-//        try {
-//            // 대표 이미지 저장
-//            if (reqDTO.getPhotoUrlFile() != null && !reqDTO.getPhotoUrlFile().isEmpty()) {
-//                String imageFilename = UUID.randomUUID() + "_" + reqDTO.getPhotoUrlFile().getOriginalFilename();
-//                Path imagePath = Paths.get(uploadDir + imageFilename);
-//                Files.write(imagePath, reqDTO.getPhotoUrlFile().getBytes());
-//                reqDTO.setPhotoUrl(imageFilename);
-//            } else {
-//            // 기존 값 유지
-//            reqDTO.setPhotoUrl(resumePS.getPhotoUrl());
-//        }
-//
-//        } catch (Exception e) {
-//            throw new Exception400("파일 업로드 실패");
-//        }
+        if (photoUrlFile != null && !photoUrlFile.isEmpty()) {
+            String uploadDir = System.getProperty("user.dir") + "/upload/";
+            try {
+                Files.createDirectories(Paths.get(uploadDir));
+                String imageFilename = UUID.randomUUID() + "_" + photoUrlFile.getOriginalFilename();
+                Path imagePath = Paths.get(uploadDir + imageFilename);
+                Files.write(imagePath, photoUrlFile.getBytes());
+
+                reqDTO.setPhotoUrl(imageFilename);
+            } catch (Exception e) {
+                throw new Exception400("파일 업로드 실패");
+            }
+        }
         
         // 3. 이력서 업데이트
         resumePS.update(reqDTO.getTitle(), reqDTO.getPhotoUrl(), reqDTO.getIsPublic(), reqDTO.getSummary(), reqDTO.getPositionType(), reqDTO.getSelfIntroduction());
@@ -211,15 +205,15 @@ public class ResumeService {
     }
 
 
-    public ResumeResponse.UpdateDTO getUpdateForm(Integer resumeId) {
+    public ResumeResponse.UpdateDTO getUpdateForm(Integer resumeId, Integer sessionUserId) {
         // 1. 해당 이력서 존재 확인
         Resume resume = resumeRepository.findByIdJoinUser(resumeId);
         if (resume == null) throw new Exception404("이력서가 존재하지 않습니다.");
 
-        // 2. (선택) 권한 확인
-        // if (!(resume.getUser().getId().equals(sessionUserId))) {
-        //     throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
-        // }
+        // 2. 권한 확인
+         if (!(resume.getUser().getId().equals(sessionUserId))) {
+             throw new Exception403("해당 이력서에 대한 권한이 없습니다.");
+         }
 
         // 3. 이력서 관련 데이터 조회
